stack command added for maptiles
Stack commands
"MAPTILES"

and

"MAPTILES lat,lon,lat,lon,zoom"

is now working.

Unsure if it is the correct way to create a stack command.

I first tried by importing the stack module into maptiles.py but it did not work. I think it is because maptiles.py is started by the client and commands are first processed by the server.